Lua class signal handlers can now return values to C.
Signals are now able to have handlers that return values to C or Lua via
the luaH_object_emit_signal C function and class:emit_signal(..)
metamethod. Signal handler execution stops after the first handler
returns a value (including nil) and the number of values  returned is
returned by the luaH_object_emit_signal C function or directly from the
class:emit_signal(..) metamethod.

If you ask for a specific number of arguments the returned number of
arguments is adjusted to match that value (either by nil padding the lua
stack or truncating the stack).

It is also possible to give LUA_MULTRET as the number of expected
arguments. If a given signal has no signal handlers or all signal
handlers returned no values then 0 is returned.

It is worth noting that if you asked for 0 return arguments signal
handling will not stop if you return a value in a handler for that
signal. If the signal is emitted from lua then the number of return
values is always LUA_MULTRET. This could be changed in the future.