Remove window from windows list before raising destroy signal & bugfix
There was a problem in the generation of the windows table. I was using
`(window = list->pdata[i]) != NULL` to iterate over the items in the
array but I did not know you can't trust g_ptr_array_remove to set
list->pdata[i] to NULL after an item is removed. Using list->len
instead.