Add silent option to soup.add_cookie function
The cookie jar changed callback doesn't emit "cookie-changed" signals
when the silent option is set.

This is the default option and when used it assumes that cookies being
added are already known to the user and there is no need check or add
them again to the cookie jar file or database.