Fixes #107: use ping cmd in unique.is_running() to confirm luakit open
Found unique_app_is_running(..) returning true on some systems when
luakit wasn't running. Don't know what is causing this and it could just
be a problem with my dbus setup however other users have reported the
same issue.

This commit modifies the unique.is_running() func to send a dummy ping
message to the supposedly running instance to check that it is indeed
open and able to receive messages.