Fix Web Inspector toggling
When inspecting a new element (via context menu or via inspect button in Web
inspector toolbar at the bottom) when web inspector is shown, it was always
changing it's attached state, i.e. when attached it got detached and vice
versa.

If we want to start attached, we should say we want start attached, not toggle
the state.