Command to run given commands in new tab
The new command `tab` takes as arguments a command and its arguments, and
creates a new tab and runs the given command and arguments in that tabs
context.

For instance, suppose you had created your own command `showTicket`, which
took the id of a ticket and opened the correct URL for it, running the command
`tab showTicket 47` would open the URL in a new tab.