Allow for prefixes on follow hints.
Sometimes it's easier to follow a link by just typing out the link text, but
this is not possible when using letters as hint labels. To get around this, we
add a transformer to the charset function pipeline that allows for a prefix to
be placed on all labels. For example, 'faa' becomes ',faa'. Now we can very
easily match using either method with no conflicts.