Reset search results after "Clear results" click
Say I do some history search, then click "clear results" button. After
that all the results are supposed to be gone. This blank white page
makes no use to the user anymore after that. So the right thing to do,
in my opinion, would be to reset the search term and show all the
history back again.