Initial port of follow.lua
In WebKit 2, there is no way to access the JavaScript contexts of
sub-frames, which makes it impossible to use the existing JS-based
follow mode implementation. The WebKitDOM API, however, provides the
access previously provided by JS. Therefore, a Lua wrapper in the form
of the dom_document and dom_element classes has been implemented, and is
used here to port follow mode to run on the WebKit Lua web process.

This is more of a reimplementation than a port; some features behave
slightly differently, some may not work perfectly yet, performance is
probably sub-optimal initially, etc.