Update webview scroll interface for webkit2
There is no way to quicky find the current scroll information on the
UI-thread; as a workaround, we listen to DOM events related to window
resizing and scrolling and document resizing on the Web process, and
send updated values to the UI thread. This method allows us to keep the
old Lua scrolling interface. Currently this seems to perform well,
even though the document height is recalculated on each DOM
modification...