Port adblock module to webkit2
Instead of listening to the navigation-request signal, which fires after
the request has begun, we listen to the send-request signal on the web
process, which fires before the request begins; the difference is vital
when blocking tracking URLs.

The rule parsing and UI remain the same; when the rule tables are
updated, they are sent to the web module over the IPC channel. This
doesn't appear to affect startup time adversely at all, since the
serialized rules table (for my lists) currently clocks in at around 4 Mb.