Allow delaying IPC message handling via type mask
Currently, the grand plan is to add support for exporting Lua functions
to JavaScript; since these functions must now run on the web process in
WebKit 2, and current uses often require returning values to JavaScript
(and not mere value passing to Lua), the only viable solution seems to
be to send an IPC message to the UI process, which executes the Lua
function, and block while waiting for a response.

In the interest of not blocking the web process for too long, and just
generally retaining our sanity, we need to ignore all other messages
while we wait for the reply containing the value to return to JS. The
last thing we need is to start loading some web module or running some
web Lua signal handler while we're blocking the whole process.

This commit supports that; unwanted messages are added to a queue and
dispatched later.