Fix webkit2 inspector
For WebKit 2, the inspector basically just works, and from the API it
doesn't seem like there's all that much you can do with it anyway. The
documentation makes some mention of getting a GtkWidget but of course
completely fails to specify how one might do so.

The iview parameter is now useless, and it's been replaced with a
boolean inspector_open property.