Fix web process connection race condition
Previously, the socket file was created and removed with each connection
from a web process. This caused web processes to abort on startup when
they attempted to connect to a socket which wasn't present on disk. This
resulted in tabs occasionally failing to load.

Now, a single socket is opened and kept open for the duration of the
program.