window: maintain constant status/input bar height (#594)
Uses the newly-wrapped stack widget. This necessitates moving from a
widget structure with two possible bar states (sbar and ibar) to one
with three (sbar, mbar, and ibar), to avoid showing and hiding the input
bar widget (necessary for the stack widget to work). Moving to this
structure is also necessary to keep the same API.

This still shows and hides the prompt text labels: this is unlikely to
cause an issue, as they're almost certain to have the same height as the
sbar labels.