Improve adblock load blocking
This change does not add a load block if the adblocking module is
disabled, since there's no need to block page loads until the adblock
module itself is loaded.

We need to conditionally unblock in a callback on the switched-page
signal, because it's possible for the module to be enabled / disabled in
between the creation of a tab and the user first switching to it.