Avoid a panic if there's problems reading /etc/hosts
In case that, for whatever reason, the file /etc/hosts is missing
from the system or it otherwise can't be read an unprotected call
to io.lines() raises an error which bubbles up to panic the host.

This seems undesireable, certainly a full panic is not wanted.

Here we look the other way and pretend the hosts file was just empty.