Fix tabs opened in background being blank
This issue was introduced by e333849; WebKit requires that tabs created
for related views (i.e. in response to the GTK "create" signal) do
not have a URL set, but that commit caused the default to be the
luakit://newtab/ page.

This commit fixes that issue by introducing an optional parameter to
prevent this behaviour. Slightly ugly, but does the trick.

Fixes #834.
Fixes #854.