Fallback to getClientRects for elements with child
In case a nested child element is present, the client_rects_api returns the bounds for the child element. If no child is present it calculates the bounding box for the current element. Passing the parent to it when a child exists does not work.

The API is deprecated since webkit2gtk 2.22. Maybe it is a good idea to switch back to the JS implementation for everything. For now, this should fix https://github.com/luakit/luakit/issues/840