Remove debug symbols from builds without DEVELOPMENT_PATHS
Debug symbols have been compiled in by default, making the resulting binary bigger and slower than it needs to be. There are two ways to solve this. Either compile with debug symbols and strip them when executing the install target (install -s ...), or not compile them in in first place.

Stripping debug symbols at install time is annoying to change, in case an actual installation with debug symbols is wanted. So I went with the latter and removed the debug symbol generation. If you really need them, add the `-ggdb` back to the `CFLAGS`.

As I assume that mostly developers want them, and as I further assume that they often build with `DEVELOPMENT_PATHS=1`. I've set the `-ggdb` flag in this case by default.