Search LUAKIT_CONFIG_PATH for rc.lua and theme.lua
When building with a non-standard XDGPREFIX luakit fails to start as it
is unable to find rc.lua and then later theme.lua if the user lacks
these files in the configured path.

Add the configured LUAKIT_CONFIG_PATH to the list of paths to look in
for a smoother experience with custom installs.