Adding a switch to turn off pdfjs rendering.
Actually, it is default false, which arguably may not match the majority's
taste.  I can be swayed.

In implementation, this adds an application.enable_pdfjs setting that is
evaluated lib/webview.lua on opening a view.  Because the pdfjs setting
is not yet exposed in libwebkitgtk, we have a C-language backend (that,
itself, depends on webkit_settings_get_all_features, which was only added
around 2.42.