Get basic Github CI working.
This adds basic coverage only -- e.g. non-luajit is not covered (and
should be), as well as other Linux distributions.

`make` and `make run-tests` are separated into separate steps so that
the build output (which is pretty verbose) can be collapsed separately.

The Github Actions builders are currently using Ubuntu 22.04 LTS, which
contains a slightly older glib without the `G_APPLICATION_DEFAULT_FLAGS`
symbol; turns out a conditional check was already there but was slightly
off, so fixing that in passing.