unique_instance: allow opening command line URLs in the current tab
For a kiosk use-case, it can be useful to be able to change the
currently open page externally. There are several options on how to
implement this but the simplest is to make use of the existing
unique_instance plugin and allow modifying its behaviour to open the
(single) URL provided by a secondary instance in the current tab.

Requested-by: puterboy <github@kosowsky.org>